requirements.txt
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.4.2
googleapis-common-protos==1.61.0
grpcio==1.59.3
idna==3.4
importlib-metadata==6.8.0
kafka-python==2.0.2
mysql-connector==2.2.9
numpy==1.26.2
opentelemetry-api==1.21.0
opentelemetry-distro==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
pandas==2.1.3
git+https://github.com/yurisa2/petrosa-utils.git
protobuf==4.25.1
py==1.11.0
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
requests==2.31.0
retry==0.9.2
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wrapt==1.16.0
zipp==3.17.0